Refactor Widget, Divider, Panel and Window widgets to own modules
